Reorganize interface source code and documentation (#2338)
* Remove tape subclasses and old QNode

* remove more files

* remove interfaces and test fixes

* Move interfaces and interface tests out of the batch folder

* rename torch gpu test

* fix codecov

* Reorganize interface source code and documentation

* fix docs

* changelog

* changelog

* bugfix

* Revert "bugfix"

This reverts commit abb1bce3c70c2669e3721f5f78133d10e6e23dbc.

* fix
